+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `transport.py` was introduced with common base class for TCP and UDP protocols, where several implementations are present
   for functions that work the same for both protocols. Some modifications, especially for `read` functions, were required for doing this.
   Other functions were implemented separately.
+- QMI drivers for Tenma 72-series power supply units in `qmi.instruments.tenma`
 
 ### Changed
 - Changed the stopping of contexts in `qmi_proc.proc_stop()` to happen in reverse order to `proc_start()`.